Simple manager blueprint single host system test fix (second test in the suite was always failing), and some refactory.
