+++ cosmo_tester/test_suites/test_simple_manager_blueprint/hello_world_singlehost_test.py
@@ -13,200 +13,25 @@
 #    * See the License for the specific language governing permissions and
 #    * limitations under the License.
 
-import os
-
-from fabric import api as fabric_api
-
-from cloudify.workflows import local
-from cloudify_cli import constants as cli_constants
-
-from cosmo_tester.framework.util import create_rest_client
 from cosmo_tester.test_suites.test_blueprints.hello_world_bash_test \
     import AbstractHelloWorldTest
-from cosmo_tester.framework.git_helper import clone
-
-MANAGER_BLUEPRINTS_REPO_URL = 'https://github.com/cloudify-cosmo/' \
-                              'cloudify-manager-blueprints.git'
+from cosmo_tester.test_suites.test_simple_manager_blueprint\
+    .abstract_single_host_test import AbstractSingleHostTest
 
 
-class HelloWorldSingleHostTest(AbstractHelloWorldTest):
+class HelloWorldSingleHostTest(AbstractHelloWorldTest, AbstractSingleHostTest):
 
     def setUp(self):
         super(HelloWorldSingleHostTest, self).setUp()
-        blueprint_path = self.copy_blueprint('openstack-start-vm')
-        self.blueprint_yaml = blueprint_path / 'blueprint.yaml'
-        self.prefix = 'simple-host-{0}'.format(self.test_id)
-        self.manager_blueprint_overrides = {}
-
-        self.inputs = {
-            'prefix': self.prefix,
-            'external_network': self.env.external_network_name,
-            'os_username': self.env.keystone_username,
-            'os_password': self.env.keystone_password,
-            'os_tenant_name': self.env.keystone_tenant_name,
-            'os_region': self.env.region,
-            'os_auth_url': self.env.keystone_url,
-            'image_id': self.env.centos_7_image_name,
-            'flavor': self.env.medium_flavor_id,
-            'key_pair_path': '{0}/{1}-keypair.pem'.format(self.workdir,
-                                                          self.prefix)
-        }
-
-        self.logger.info('initialize local env for running the '
-                         'blueprint that starts a vm')
-        self.local_env = local.init_env(
-            self.blueprint_yaml,
-            inputs=self.inputs,
-            name=self._testMethodName,
-            ignored_modules=cli_constants.IGNORED_LOCAL_WORKFLOW_MODULES)
-
-        self.logger.info('starting vm to serve as the management vm')
-        self.local_env.execute('install',
-                               task_retries=10,
-                               task_retry_interval=30)
-        self.public_ip_address = \
-            self.local_env.outputs()['simple_vm_public_ip_address']
-        self.private_ip_address = \
-            self.local_env.outputs()['simple_vm_private_ip_address']
-
-        self.addCleanup(self.cleanup)
+        self.setup_simple_manager_env()
 
     def test_hello_world_singlehost(self):
-        remote_manager_key_path = '/home/{0}/manager_key.pem'.format(
-            self.env.centos_7_image_user)
-        self.bootstrap_simple_manager_blueprint(remote_manager_key_path)
-        inputs = {
-            'server_ip': self.public_ip_address,
-            'agent_user': self.env.centos_7_image_user,
-            'agent_private_key_path': remote_manager_key_path
-        }
-        self._run(blueprint_file='singlehost-blueprint.yaml', inputs=inputs)
-
-    def _setup_external_components_vm(self):
-        blueprint_path = self.copy_blueprint('external-components-vm')
-        self.blueprint_yaml = \
-            blueprint_path / 'external-components-blueprint.yaml'
-        self.prefix = 'external-components-host-{0}'.format(self.test_id)
-        self.manager_blueprint_overrides = {}
-
-        self.ext_inputs = {
-            'prefix': self.prefix,
-            'external_network': self.env.external_network_name,
-            'os_username': self.env.keystone_username,
-            'os_password': self.env.keystone_password,
-            'os_tenant_name': self.env.keystone_tenant_name,
-            'os_region': self.env.region,
-            'os_auth_url': self.env.keystone_url,
-            'image_id': self.env.centos_7_image_name,
-            'flavor': self.env.medium_flavor_id,
-            'key_pair_path': '{0}/{1}-keypair.pem'.format(self.workdir,
-                                                          self.prefix)
-        }
-
-        self.logger.info('initialize external '
-                         'components local env for running the '
-                         'blueprint that starts a vm of es and influx')
-        self.ext_local_env = local.init_env(
-            self.blueprint_yaml,
-            inputs=self.ext_inputs,
-            name=self._testMethodName,
-            ignored_modules=cli_constants.IGNORED_LOCAL_WORKFLOW_MODULES)
-
-        self.logger.info('starting vm to serve as the management vm')
-        self.ext_local_env.execute('install',
-                                   task_retries=10,
-                                   task_retry_interval=30)
-        self.external_components_public_ip = \
-            self.ext_local_env.outputs()[
-                'external_components_vm_public_ip_address']
-        self.external_components_private_ip = \
-            self.ext_local_env.outputs()[
-                'external_components_vm_private_ip_address']
-        self.addCleanup(self.cleanup_ext)
-
-    def test_external_components(self):
-        self._setup_external_components_vm()
-        remote_manager_key_path = '/home/{0}/manager_key.pem'.format(
-            self.env.centos_7_image_user)
-        additional_bootstrap_inputs = {
-            'elasticsearch_endpoint_ip': self.external_components_public_ip,
-            'influxdb_endpoint_ip': self.external_components_public_ip
-        }
-        self.logger.info(str(additional_bootstrap_inputs))
-        self.bootstrap_simple_manager_blueprint(
-            remote_manager_key_path, additional_bootstrap_inputs)
-        inputs = {
-            'server_ip': self.public_ip_address,
-            'agent_user': self.env.centos_7_image_user,
-            'agent_private_key_path': remote_manager_key_path
-        }
+        self.bootstrap_simple_manager_blueprint()
         self._run(blueprint_file='singlehost-blueprint.yaml',
-                  inputs=inputs,
-                  influx_host_ip=self.external_components_public_ip)
+                  inputs=dict(self.access_credentials,
+                              **{'server_ip': self.public_ip_address}))
 
     def _do_post_uninstall_assertions(self, context):
         instances = self.client.node_instances.list(self.test_id)
         for x in instances:
             self.assertEqual('deleted', x.state)
-
-    def _bootstrap(self):
-        self.cfy.bootstrap(blueprint_path=self.test_manager_blueprint_path,
-                           inputs_file=self.test_inputs_path,
-                           task_retries=5)
-        self.addCleanup(self.cfy.teardown)
-
-    def bootstrap_simple_manager_blueprint(self, remote_manager_key_path,
-                                           additional_bootstrap_inputs=None):
-        self.manager_blueprints_repo_dir = clone(MANAGER_BLUEPRINTS_REPO_URL,
-                                                 self.workdir)
-        self.test_manager_blueprint_path = \
-            os.path.join(self.manager_blueprints_repo_dir,
-                         'new', 'simple-manager-blueprint.yaml')
-
-        # using the updated handler configuration blueprint to update the
-        # package urls in the simple manager blueprint
-        # self._update_manager_blueprint()
-
-        self.bootstrap_inputs = {
-            'public_ip': self.public_ip_address,
-            'private_ip': self.private_ip_address,
-            'ssh_user': self.env.centos_7_image_user,
-            'ssh_key_filename': self.inputs['key_pair_path'],
-            'agents_user': self.env.centos_7_image_user,
-            'resources_prefix': '',
-        }
-        self.bootstrap_inputs.update(additional_bootstrap_inputs or {})
-
-        # preparing inputs file for bootstrap
-        self.test_inputs_path = \
-            self.cfy._get_inputs_in_temp_file(self.bootstrap_inputs,
-                                              self._testMethodName)
-        self._bootstrap()
-        self._running_env_setup(self.public_ip_address)
-
-        self.logger.info('Uploading key file to manager...')
-        with fabric_api.settings(host_string=self.public_ip_address,
-                                 user=self.env.centos_7_image_user,
-                                 key_filename=self.inputs['key_pair_path']):
-            fabric_api.put(self.inputs['key_pair_path'],
-                           remote_manager_key_path)
-
-    def _running_env_setup(self, management_ip):
-        self.env.management_ip = management_ip
-        self.client = create_rest_client(management_ip)
-        response = self.client.manager.get_status()
-        if not response['status'] == 'running':
-            raise RuntimeError('Manager at {0} is not running.'
-                               .format(management_ip))
-
-    def cleanup(self):
-        self.local_env.execute('uninstall',
-                               task_retries=40,
-                               task_retry_interval=30)
-        self.env.management_ip = None
-
-    def cleanup_ext(self):
-        self.ext_local_env.execute('uninstall',
-                                   task_retries=40,
-                                   task_retry_interval=30)
-        self.env.management_ip = None
